components

+++ db/migrate/20210926171632_create_services.rb
@@ -2,10 +2,10 @@ class CreateServices < ActiveRecord::Migration[6.1]
   def change
     create_table :services do |t|
       t.string :job_type
-      t.integer :cost
+      t.float :cost
       t.string :date
       t.belongs_to :worker, null: false, foreign_key: true
-     
+
       t.timestamps
     end
   end
